+++ code/src/cljs/sixsq/nuvla/ui/session/utils.cljs
@@ -3,7 +3,6 @@
     [clojure.string :as str]))
 
 (def ^:const user-tmpl-email-password "user-template/email-password")
-(def ^:const user-tmpl-email-invitation "user-template/email-invitation")
 (def ^:const session-tmpl-password "session-template/password")
 (def ^:const session-tmpl-password-reset "session-template/password-reset")